Fix conversion warnings

+++ src/supertux/moving_object.cpp
@@ -104,7 +104,7 @@ bool
 MovingObject::is_within_bounds(const Rectf& cliprect) const
 {
   const Rect& bbox = get_bbox().to_rect();
-  auto grow_amount = std::max(bbox.get_width(), bbox.get_height());
+  auto grow_amount = static_cast<float>(std::max(bbox.get_width(), bbox.get_height()));
   return cliprect.grown(grow_amount).contains(bbox);
 }
